papyr/requirements.txt
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.1
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.28.2
PyJWT==2.8.0
pymongo==4.7.1
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
Werkzeug==3.0.2
wsproto==1.2.0
zope.event==5.0
zope.interface==6.3
